Add tests for CLI interface
