+++ tests/conftest.py
@@ -1,0 +1,1 @@
+pytest_plugins = ["pytester"]
